backend-django/requirements.txt
Django>=5.1.3
djangorestframework>=3.16.1
django-cors-headers>=4.3
requests>=2.31
python-dotenv>=1.0
